move Resume continent components into their own Resume folder
